# specs/001-deep-research-agent/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Deep Research Agent API
  description: |
    REST API for the Deep Research Agent. Provides endpoints for chat management,
    research queries, and user preferences. Compatible with Databricks chat agent patterns.
  version: 1.0.0
  contact:
    name: Deep Research Agent Team
servers:
  - url: /api/v1
    description: API v1
security:
  - databricksAuth: []
...
paths:
...
  /preferences:
    get:
      tags: [Preferences]
      operationId: getPreferences
      summary: Get user preferences
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Preferences]
      operationId: updatePreferences
      summary: Update user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    databricksAuth:
      type: http
      scheme: bearer
      description: Databricks workspace OAuth token
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    UserPreferences:
      type: object
      properties:
        system_instructions:
          type: string
          nullable: true
          maxLength: 10000
        default_depth:
          type: string
          enum: [auto, light, medium, extended]
        ui_preferences:
          type: object
          additionalProperties: true
        updated_at:
          type: string
          format: date-time
    UpdatePreferencesRequest:
      type: object
      properties:
        system_instructions:
          type: string
          maxLength: 10000
        default_depth:
          type: string
          enum: [auto, light, medium, extended]
        ui_preferences:
          type: object
          additionalProperties: true

    # ============================================================================
    # ERROR SCHEMA
    # ============================================================================
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true